Fix 'd' problem on translation
Since in spanish you use 'de' to say 'of' it gets mistaken for a d in the date format. An easy way to fix this is to use 'De' with capital D so it doesn't get mistaken. It's better to have a misplaced capital letter to a useless date name (the dates get rendered as "Domingo 22 22e Junio 22e 2013" instead of "Domingo 22 de Junio de 2013"