Fix Drop Down Selector Font Size
Month and Year drop down selectors had the font size set too high. This
was causing portions of the month and year names to render underneath
the selector's chrome. Making the fonts slightly smaller resolves the
issue. Please see the associated ticket for this change to see before
and after screen captures.