Update picker.date.js
1. removed invalid character in line 581
2. changed determination of 'fromValue' flag on init.

When having a node like this:

     <input value="2014-04-10" data-value="2014-04-10" />

And initializing pickadate like: 

     $("input").pickadate( {formatSubmit: "yyyy-mm-dd"});

The *expected* behaviour is to set the date to April 10, 2014. 

The previous implementation took the value from data-value but determined the fromValue flag by evaluating !!elementValue. This means, as long as value-attribute is set, the data-value attribute is treated with zero-based month-index, so the date would be set to March 10, 2014.