Explicitly define ignore files
https://github.com/amsul/pickadate.js/issues/432

Bower is not installing a blank folder for pickaday in some environments
(Heroku), which appears to be related to the way in which the `ignore`
section of the package manifest is defined. It's not clear from the
bower documentation how conflicts might be resolved (for example, * and
"not lib" would still ignore lib, which is what's happening in my
environment now, lib is never being downloaded).

Removed the `*` declaration and explicitly defined paths to ignore in
the `ignore` section of `package.json`.