Set pointer capture on pointerdown on picker input
Chrome 73 starts sending click even to the common
ancestor of down and up events even when one was
inside the input element and the other one was
outside. This caused this library to fail in the
following scenario:
User presses down on the picker input which causes
the date picker and overlay to be shown. When user
releases the button the overlay gets the up event
and the common ancestor of these elements will get
the click. Since there is a click handler on the
document to close the popup, the calendar popup immediately
closes.

Using pointercapture API to capture the rest of the
pointerevents stream to the picker input it ensures
the click event target is the picker input and that ensures
the click handler doesn't dismiss the popup.