Removed setTimeOut in stop method
setTimeOut for unbinding in stop method has been removed, due to the fact it was provoking errors when trying to reinitialize the datepicker. Re-initialize is to call the method stop() and then start() or datepicker. The result was having the datepicker initialized but nothing happened when clicking on it (because the events were unbinding by the previous close call)